Use github.com/myitcv/docker-compose (#14)
Using github.com/myitcv/docker-compose simplifies the use of
docker-compose in this project, but also enables other projects to
compose this docker-compose.yml into their own setups.

As part of this we ensure that the _scripts/setup.sh script can be run
in the context of another project. The other project does that by
setting COMPOSE_FILE to the "environment" of compose config files, a
list that can be extended by any call to docker-compose using the -f
flag.