Test in a Go test
Testing in a regular Go test is more robust than asking someone to run a
set of bash commands.

Add a setup command to replace the existing bash script.

Remove the wait command as a precursor to further changes to create a
server command, making gitea_prestep an HTTP endpoint.