cmd/gitea: move to create unique user and repos underneath that
Prior to hosting our own gitea instance, we were limited by the fact
that repos we created were made public under github.com/userguides/.
This meant that we had to name the repos unique things like
helloabcd1234. This is very awkward to read. Far better would be to have
unique usernames and then repos underneath that.

Now that we are hosting our own gitea instance we can do just that, and
this commit makes that change.