cmd/gitea: serve generated repo as fully qualified path (#64)
Currently we only serve the generated repo as the base of the path, e.g.
if the repo is gopher.live/u1234/hello, we return REPO1=hello. This is
not much help, because we are, in the case on no random part, simply
echoing back to the caller the pattern they specified. Furthermore, they
also need to construct the fully qualified path to the repo.

Fix that by returning REPO1=gopher.live/u1234/hello. This also means
that each environment variable returned has some random part to it,
which makes much more sense when we consider the sanitise step.